code refactor split into files
